DocSearchApi/requirements.txt
fastapi
langchain==0.0.313
uvicorn[standard]
openai==0.28.1
tiktoken==0.5.1
python-dotenv
azure-search-documents==11.4.0b8
azure-identity==1.13.0b4
